requirements.txt
aioredis==1.3.1
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
billiard==3.6.3.0
black==20.8b1
boto3==1.16.59
botocore==1.19.59
celery==5.0.5
certifi==2020.11.8
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.24.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.3
Django==3.1.5
django-channels-jwt-auth-middleware==1.0.0
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
msgpack==1.0.2
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.7
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.14
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
redis==3.5.3
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.2.0
